LayoutTests/html5lib/resources/tables01.dat

#data
<table><th>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<table><td>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<table><col foo='bar'>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <col>
|           foo="bar"

#data
<table><colgroup></html>foo
#errors
#document
| <html>
|   <head>
|   <body>
|     "foo"
|     <table>
|       <colgroup>
